+++ CHANGELOG.md
@@ -1,6 +1,24 @@
 # Changelog
 All Notable changes to `infoconnect-php-client` will be documented in this file
 
+## 0.2.0 - 2015-1-17
+
+### Added
+- Support for Companies POST Search api method
+- Support for Companies POST Count api method
+
+### Deprecated
+- Nothing
+
+### Fixed
+- Nothing
+
+### Removed
+- Nothing
+
+### Security
+- Nothing
+
 ## 0.1.0 - 2015-12-19
 
 ### Added